file upload and download

+++ controllers/admin.js
@@ -1,5 +1,7 @@
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator/check');
 const Product = require('../models/product');
+
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
@@ -14,9 +16,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.image;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            docTitle: 'Add Product', 
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image!',
+            validationErrors: []
+        });
+    }
     const errors = validationResult(req);
     if ( !errors.isEmpty() ) {
         return res.status(422).render('admin/edit-product', {
@@ -26,7 +43,6 @@ exports.postAddProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description
             },
@@ -34,6 +50,7 @@ exports.postAddProduct = (req, res, next) => {
             validationErrors: errors.array()
         });
     }
+    const imageUrl = image.path;
     const product = new Product({
         title: title,
         imageUrl: imageUrl, 
@@ -48,21 +65,6 @@ exports.postAddProduct = (req, res, next) => {
         res.redirect('/admin/products');
     })
     .catch(err => {
-        // return res.status(500).render('admin/edit-product', {
-        //     docTitle: 'Add Product', 
-        //     path: '/admin/add-product',
-        //     editing: false,
-        //     hasError: true,
-        //     product: {
-        //         title: title,
-        //         imageUrl: imageUrl,
-        //         price: price,
-        //         description: description
-        //     },
-        //     errorMessage: 'Database operation failed, please try again!',
-        //     validationErrors: []
-        // });
-        //res.redirect('/500');
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
@@ -101,7 +103,7 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDescription = req.body.description;
     const errors = validationResult(req);
     if ( !errors.isEmpty() ) {
@@ -112,7 +114,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId
@@ -129,7 +130,10 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDescription;
-        product.imageUrl = updatedImageUrl;
+        if (image) {
+            fileHelper.deleteFile(product.imageUrl);
+            product.imageUrl = image.path;
+        }
         return product.save()
         .then(result => {
             console.log('UPDATED PRODUCT');
@@ -162,7 +166,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({_id: prodId, userId: req.user._id});
+        })
         .then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
